requirements.txt
fastapi
uvicorn
yt-dlp
spotipy
python-dotenv
